semantic/requirements.txt
chromadb==0.4.24
sentence-transformers==2.5.1
scikit-learn==1.6.1
openai==1.14.0
lxml==5.1.0
pyyaml==6.0.1
python-dotenv==1.0.1
llama-cpp-python==0.2.56
numpy==1.26.4
pandas==2.2.1
tqdm==4.66.2
torch==2.2.1
transformers==4.38.2
requests==2.31.0 